+++ scripts/protoc-swagger-gen.sh
@@ -16,10 +16,10 @@ for dir in $proto_dirs; do
   fi
 done
 
-cd ../third_party/proto
 echo "Generate cosmos swagger files"
 
-proto_dirs=$(find ./cosmos ./ethermint ./ibc -path -prune -o -name '*.proto' -print0 | xargs -0 -n1 dirname | sort | uniq)
+proto_dir="../third_party/proto"
+proto_dirs=$(find "${proto_dir}/ethermint" "${proto_dir}/ibc" -path -prune -o -name '*.proto' -print0 | xargs -0 -n1 dirname | sort | uniq)
 for dir in $proto_dirs; do
   # generate swagger files (filter query files)
   query_file=$(find "${dir}" -maxdepth 1 \( -name 'query.proto' -o -name 'service.proto' \))
@@ -29,7 +29,9 @@ for dir in $proto_dirs; do
   fi
 done
 
-cd ../..
+buf generate --template buf.gen.swagger.yaml "buf.build/cosmos/cosmos-sdk:8cb30a2c4de74dc9bd8d260b1e75e176"
+
+cd ..
 
 echo "Combine swagger files"
 # combine swagger files